refactor(pipeline): reorganize tile stages into tiles directory
Rename src/stages to src/tiles to reflect domain-driven packaging. All tile-related modules (fetch, merge, convert, validate, index-gen) are now colocated in a semantically named directory, improving codebase navigation and domain clarity.
