+++ apps/pipeline/src/pipeline.ts
@@ -1,9 +1,9 @@
 import { EXIT_CODES, PATHS, UPSTREAM } from '@/config.ts';
 import type { PipelineLogger } from '@/shared/logger.ts';
-import { executeFetch, getCommitHash, parseYearsFromDirectory } from '@/stages/fetch.ts';
-import { runIndexGenStage } from '@/stages/index-gen.ts';
 import { PipelineCheckpoint } from '@/state/checkpoint.ts';
 import { acquireLock, registerCleanupHandlers, releaseLock } from '@/state/lock.ts';
+import { executeFetch, getCommitHash, parseYearsFromDirectory } from '@/tiles/fetch.ts';
+import { runIndexGenStage } from '@/tiles/index-gen.ts';
 import type { ValidationResult } from '@/types/pipeline.ts';
 import { generateReport } from '@/validation/report.ts';
 import { YearProcessor } from '@/year-processor.ts';
