+++ apps/pipeline/src/tiles/index-gen.test.ts
@@ -3,7 +3,7 @@ import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import type { PipelineLogger } from '@/shared/logger.ts';
-import { generateYearIndex } from '@/stages/index-gen.ts';
+import { generateYearIndex } from '@/tiles/index-gen.ts';
 
 function createMockLogger(): PipelineLogger {
   return {
