+++ apps/pipeline/src/tiles/validate.test.ts
@@ -4,7 +4,7 @@ import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import type { PipelineLogger } from '@/shared/logger.ts';
-import { runValidateForYear } from '@/stages/validate.ts';
+import { runValidateForYear } from '@/tiles/validate.ts';
 
 const FIXTURES = path.join(process.cwd(), 'tests', 'fixtures');
